requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
faker>=8.0.0,<8.1.0
flake8>=3.6.0,<3.7.0
django-rest-swagger>=2.0.0,<2.1.0